Move reader and writer to separate files
